update the query names, and moved the searchUserCard file to common folder
