+++ server/ExpertBridge.Core/Requests/SearchPostRequest.cs
@@ -2,6 +2,6 @@ namespace ExpertBridge.Core.Requests;
 
 public class SearchPostRequest
 {
-    public string query { get; set; }
-    public int? limit { get; set; }
+    public string Query { get; set; }
+    public int? Limit { get; set; }
 }
